chore(gatsby-cli): Convert yurnalist logger to TypeScript
