+++ packages/gatsby-cli/src/reporter/loggers/yurnalist/index.ts
@@ -1,0 +1,131 @@
+import { onLogAction } from "../../redux"
+import {
+  Actions,
+  LogLevels,
+  ActivityLogLevels,
+  ActivityTypes,
+} from "../../constants"
+
+import { createReporter } from "yurnalist"
+import ProgressBar from "progress"
+import chalk from "chalk"
+import { IUpdateActivity } from "../../redux/types"
+
+type YurnalistActivities = {
+  [activityId: string]: {
+    text: string | undefined
+    statusText: string | undefined
+    update(payload: IUpdateActivity["payload"]): void
+    end(): void
+  }
+}
+
+export function initializeYurnalistLogger() {
+  const activities: YurnalistActivities = {}
+  const yurnalist = createReporter({ emoji: true, verbose: true })
+
+  const levelToYurnalist = {
+    [LogLevels.Log]: yurnalist.log.bind(yurnalist),
+    [LogLevels.Warning]: yurnalist.warn.bind(yurnalist),
+    [LogLevels.Error]: yurnalist.error.bind(yurnalist),
+    [LogLevels.Info]: yurnalist.info.bind(yurnalist),
+    [LogLevels.Success]: yurnalist.success.bind(yurnalist),
+    [ActivityLogLevels.Success]: yurnalist.success.bind(yurnalist),
+    [ActivityLogLevels.Failed]: (text: string): void => {
+      yurnalist.log(`${chalk.red(`failed`)} ${text}`)
+    },
+    [ActivityLogLevels.Interrupted]: (text: string): void => {
+      yurnalist.log(`${chalk.gray(`not finished`)} ${text}`)
+    },
+  }
+
+  onLogAction(action => {
+    switch (action.type) {
+      case Actions.Log: {
+        const yurnalistMethod = levelToYurnalist[action.payload.level]
+        if (!yurnalistMethod) {
+          process.stdout.write(`NO "${action.payload.level}" method\n`)
+        } else {
+          let message = action.payload.text
+          if (action.payload.duration) {
+            message += ` - ${action.payload.duration.toFixed(3)}s`
+          }
+          if (action.payload.statusText) {
+            message += ` - ${action.payload.statusText}`
+          }
+          yurnalistMethod(message)
+        }
+        break
+      }
+      case Actions.StartActivity: {
+        if (action.payload.type === ActivityTypes.Spinner) {
+          const spinner = yurnalist.activity()
+          spinner.tick(action.payload.text)
+
+          const activity = {
+            text: action.payload.text,
+            statusText: action.payload.statusText,
+            update: payload => {
+              if (payload.text) {
+                activity.text = payload.text
+              }
+              if (payload.statusText) {
+                activity.statusText = payload.statusText
+              }
+
+              let message = activity.text
+              if (activity.statusText) {
+                message += ` - ${activity.statusText}`
+              }
+
+              message += ` id:"${action.payload.id}"`
+
+              spinner.tick(message)
+            },
+            end: () => {
+              spinner.end()
+            },
+          }
+          activities[action.payload.id] = activity
+        } else if (action.payload.type === ActivityTypes.Progress) {
+          const bar = new ProgressBar(
+            ` [:bar] :current/:total :elapsed s :percent ${action.payload.text}`,
+            {
+              total: action.payload.total,
+              curr: action.payload.current,
+              width: 30,
+              clear: true,
+            }
+          )
+
+          activities[action.payload.id] = {
+            text: undefined,
+            statusText: undefined,
+            update: payload => {
+              if (payload.total) {
+                bar.total = payload.total
+              }
+              if (payload.current) {
+                bar.curr = payload.current
+              }
+
+              bar.tick(0)
+            },
+            end: () => {},
+          }
+        }
+        break
+      }
+      case Actions.UpdateActivity: {
+        const activity = activities[action.payload.id]
+        activity?.update(action.payload)
+        break
+      }
+      case Actions.EndActivity: {
+        const activity = activities[action.payload.id]
+        activity?.end()
+        delete activities[action.payload.id]
+      }
+    }
+  })
+}
